backend: Use PKRelatedFields instead of nested serializers
The earlier approach in serializing the Project and Publication objects
was to have the related author fields as nested serializations. This
approach was wrong, as it created problems when trying to create/update
the objects, since the nested serializers required the entire objects.
Now we use the PrimaryKeyRelatedField, as recommended by the docs, and
manually populate the nested related fields during serialization.